api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pika==1.3.2
psycopg2-binary==2.9.9
pydantic==2.5.0